Move ReceiptUploading Components to directory
 This refactoring should make it easier to swap out whatever is being
 used for file uploading
