+++ src/components/receiptUpload/ReceiptUpload.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import ACLUCloudWok from './ACLUCloudwok';
+import Instructions from './Instructions';
+
+export default () => 
+  <div> 
+    <Instructions/>
+    <ACLUCloudWok />
+  </div>
